Fix top border for the plan and connection section
