+++ projects/packages/my-jetpack/_inc/components/my-jetpack-tab-panel/tab-content.tsx
@@ -1,5 +1,5 @@
 import { HelpContent } from './help/content';
-import { OverviewContent } from './overview-content';
+import { OverviewContent } from './overview/content';
 import { ProductsContent } from './products-content';
 import styles from './styles.module.scss';
 import { MyJetpackSection } from './types';
